Use size_t to fix comparision using RISC-V
The RISC-V compiler complains about missmatching types otherwise:
src/esphome/components/ota/ota_component.cpp: In member function 'void esphome::ota::OTAComponent::handle_()':
src/esphome/components/ota/ota_component.cpp:271:62: error: no matching function for call to 'min(unsigned int, uint32_t)'
     size_t requested = std::min(sizeof(buf), ota_size - total);
                                                              ^

+++ esphome/components/ota/ota_component.cpp
@@ -109,11 +109,11 @@ void OTAComponent::loop() {
 void OTAComponent::handle_() {
   OTAResponseTypes error_code = OTA_RESPONSE_ERROR_UNKNOWN;
   bool update_started = false;
-  uint32_t total = 0;
+  size_t total = 0;
   uint32_t last_progress = 0;
   uint8_t buf[1024];
   char *sbuf = reinterpret_cast<char *>(buf);
-  uint32_t ota_size;
+  size_t ota_size;
   uint8_t ota_features;
   std::unique_ptr<OTABackend> backend;
   (void) ota_features;
